code cleanup
- moved the presets handling to it own file
- footer is now defined in shaders
- fixed minor
- made some code parts easier
- moved code parts
- commented code
