+++ _PoiyomiToonShader/Editor/PoiPresetHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d8891d44950203948bd04b29ba39ed9e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
